qcacld-3.0: Fix STA including incorrect SMPS value in the HT caps of the reassoc req
qcacld-2.0 to qcacld-3.0 propagation

When creating the reassoc session, HT SMPS value should be copied from
existing session to the reassoc session but the session supported NSS should
be set to default value of true and not copied from exisiting session.
Session supported NSS should be the intersection of NSS supported by STA
and AP that STA will reassoc to.

CRs-Fixed: 991042
Change-Id: Ic1577882aaa9c1923ba1dac2126e0cf11e542e36
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>