qcacld-3.0: Fix incorrect handling of encrypted auth managment frame
Fix incorrect handling of encrypted auth managment frame in case of
WEP shared encryption mode.

Authentication frame length is populated based on the auth transaction
sequence number. However this will not hold good for encrypted frame
from which we will not be able to decode the transaction sequence number.
Make use of the flag passed to lim_send_auth_mgmt_frame to decide if
the frame is encrypted or not and take necessary action.

Change-Id: I61df6a3e871965cd35128fe11925855d37bf81b9
CRs-Fixed: 999599
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>