qcacld-3.0: Save the requested scan and fw mode config
Save the requested scan and fw mode config when these
new configs are sent to the firmware for update. This
saved parameter updates the current scan and fw mode
config respectively on getting the response from the
firmware. If this parameter is not saved and updated
in the current scan and fw mode configs, the next
attempt to fetch the current scan and fw mode config
provides the wrong config values.

Change-Id: Iae10d16b29038a8e321ef74ba591b032eedbe425
CRs-Fixed: 1007653
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>