qcacld-3.0: Correct the p2p scan parameters when SAP is beaconing
Correct the p2p scan's active dwell time, passive dwell time and
burst duration when a SAP session is already active. Since the
max duration of CTS-to-self is 32ms, the dwell time for p2p scans
are calculated in such a way that it does not cross 32ms if SAP
is running. This is in line with the existing STA scan parameters
when SAP is active.

Change-Id: I8089e01025ba07cc18684c277c36bb1694264214
CRs-Fixed: 998555
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>