qcacld-3.0: Start ap adapter from __wlan_hdd_add_virtual_intf
With SAP+STA concurrency there shall be a new interface created softap0 on
which SAP comes up. In the present scenario the support for starting
the ap adapter only via the change interface.
Since the SAP adapter is not started the start_bss request from the
upper layer fails.

To migitate the issue add support for the start ap adapter from the
 __wlan_hdd_add_virtual_intf.

Change-Id: I4d112441b5aa7ff91dafa9e43f9aceb18d0e13f0
CRs-Fixed: 1056331
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>