qcacld-3.0: Update the net device setup for monitor mode
Netdevice setup for the Monitor mode/Device mode are different
with Dynamic mode change the adapter is allocated for normal mode
and the device setup is done according to the normal modes.
Once the device mode change request for monitor mode comes via con_mode
the driver shall change to monitor mode, when the mode switch happens
presently driver is using the same setup as station in monitor mode.

Update the net device setup for the monitor mode once the device changes
to monitor mode.

Change-Id: I38ff0e31b80f8e800a37c10431902ab092b627ba
CRs-Fixed: 1064063
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>