qcacld-3.0: Implement vendor command to avoid freq in SAP
qcacld-2.0 to qcacld-3.0 propagation

Add support for vendor command which ask the driver
to restart the SAP if SAP is on unsafe channel.
QCA_NL80211_VENDOR_SUBCMD_AVOID_FREQUENCY is used to send
driver to restart SAP device if SAP is on any unsafe channels.
Upon receiving this command, driver will restart the SAP device
if SAP is operating on any of one the unsafe channels.

Change-Id: I187a6e0c9f3eaa96c740e50c332a7aba5796dc74
CRs-Fixed: 998048
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>