qcacld-3.0: Remove "sample" WCNSS_qcom_cfg.ini
Currently the qcacld-3.0 project contains a "sample" version of the
WCNSS_qcom_cfg.ini configuration file.  This differs from the "real"
versions that are kept on a per-product basis in product-specific
projects.  This has become a source of confusion since developers are
surprised when their modifications to the sample aren't reflected on
the product.  Remove the confusion by removing the sample file.  Folks
will then know they need to modify the product-specific files in their
product-specific projects.

Change-Id: Ia1b04ac431d3ae0e91d597bdfa6b8acfcd19d9e7
CRs-Fixed: 1065266
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>