qcacld-3.0: Purge old scan entries if max BSS limit reached
qcacld-2.0 to qcacld-3.0 propagation

In the driver, while processing the scan results, if the max
BSS limit is reached, all the newest scan results are dropped
and stale scan results are updated to upper layers.
To address this, remove the oldest scan entry if the max BSS
limit is reached.

Change-Id: Ib85a8a3b66fbf903311002887a3c5c3d10bfbccf
CRs-Fixed: 991417
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>