qcacld-3.0: Append EC IE and Interworking IE to beacon
qcacld-2.0 to qcacld-3.0 propagation

Supplicant is sending Extended capabilities (EC) IE and
Interworking IE as part of beacon IEs to the driver but
the driver is not looking for these IEs when populating
the beacon. To fix this append the EC IE and Interworking
IE to the beacon template.

Change-Id: I6e19bfacb6f83526ce80d59cf43c23c6e2a77233
CRs-Fixed: 964594
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>