qcacmn: Fix issue in green AP feature
Green AP feature for WIN allows the AP to go into power save mode
and leads to only a single chain being active.

In the current code, it is seen that the WMI API for enabling Green
AP feature is under "FEATURE_GREEN_AP" flag that is not defined for
WIN. This leads to trigger of dummy API that does no operation.

Replace "FEATURE_GREEN_AP" with proper flag "ATH_SUPPORT_GREEN_AP"
defined on WIN to enable Green AP feature.

Change-Id: I344e74c3753f1ef3b6d5d5a2bb1086c84f2ab84c
CRs-Fixed: 1074036
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>