qcacld-3.0: Use system time instead of jiffies for BSS received time
qcacld-2.0 to qcacld-3.0 propagation

During late suspend jiffies will not be incremented. Because of this
scan results are not age out as the delta of current time and the BSS
received time is not correct.

To address this, use the system time instead of jiffies for the
BSS received time and also make sure to use system time in all
other functions.

CRs-Fixed: 1018460
Change-Id: Ia14b84f1039dc4ffec6ad63550fed21e932f8012
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>