qcacld-3.0: HDD: Fix incorrect use of "\" line continuation
Currently there are instances of the "\" line continuation character
being used inside quoted strings. This has the undesired side effect
of including the indentation of the continued line as part of the
quoted string. In addition per the Linux Coding Style user-visible
strings should not be split. So remove all incorrect instances of "\"
line continuation from HDD.

Change-Id: I9c63be20f01cb14c615b0a1c786ed06a3c2094aa
CRs-Fixed: 1087272
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>