qcacld-3.0: Avoid decrementing no of connected peer twice on failure
In case rate set of the peer in assoc req doesn't match the number
of connected peers is decremented and sta is deleted. Now as part
of send del sta again the number of connected peers is decremented.
This sets the number of peer connected to 255 when last peer departs.
And thus new peer always fails to join as the no of connected peer
is always more than the allowed limit.

In this case the sta is not yet added and thus there is no need
to call del sta.

To fix this set delete_sta false while calling lim_reject_association
and thus avoid calling del sta. Also clean up the sta_ds from
hash table.

Change-Id: Ia7d837cfbbf9ddc15230aaeeca1a72acf48ac232
CRs-Fixed: 1108820
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>