qcacld-3.0: Register Bus bandwidth timer directly with kernel
Currently, bus bw timer is defined globally and it access the
adapter nodes.The crash can happen when the adapter is getting
deregistered in one thread and in another thread bus bw timer
handler is trying to access the same adapter.

To mitigate this issue, bus bw timer needs to directly registered
with kernel and once the timer callback is executed then it needs
to schedule a workqueue where all the processing is done.
While closing the adapter, do flush_work for this timer to
make sure that workqueue is done and then it is closing the adapter.

CRs-Fixed: 1104615
Change-Id: I81a7bfa43e3fb43488966dda702c6f394f9b40c7
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>