qcacld-3.0: Add framework to deprecate Wireless Extensions
Although the wlan driver currently utilizes nl80211/cfg80211 as its
primary configuration interface, it still contains support for the
legacy wireless extensions ioctl interface. This interface, however,
is not being maintained and at some point in the future it should be
removed. But before it can be removed all applications which depend
upon this interface must be converted to use other supported
interfaces such as nl80211. To aid in this effort add a framework for
detecting and logging when wireless extensions ioctls are being used.

Change-Id: Ib2f3b64f74e673f3857f14ae12ff372a9f60de8d
CRs-Fixed: 2004194
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>