qcacld-3.0: Pass txrx context durring pktlog deinit
pktlog deinit should not use cds_get_context.
Using cds_get_context for the txrx context was
causing the device to crash on the next insmod
because the deinit is happenning durring the deinitialization
of the txrx context after it has allready been unlinked from
as a global context.

Change-Id: I487951a6dd50680884223e7027e5bce1c8ddfa07
CRs-Fixed: 2004404
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>