qcacld-3.0: update hw/sw info afer modules get closed
As part of stop_modules hif_ctx  will be freed and re-allocated
as part of start_modules. So the target_name pointer in the hdd_ctx
points to a invalid memory.

Mitigate this issue by updating the target_name,hw/sw info during
when the driver_status moves to opened state.

CRs-Fixed: 2009308
Change-Id: I73d2f7537e1ff2a4dd735e1af85473b64c2bf1ae
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>