qcacld-3.0: Support fast transition with open security
wlan-cld3.driver.lnx.1.1-dev to wlan-cld3.driver.lnx.2.0-dev propagation

If station is connected to an AP with open security, fast
transition would fail as preauth is attempted only if FT
ies are present.

To fix this, Preauth is attempted even if FT ies are not
present for a 11R connection in open security.

Change-Id: I0ac15fc63b4b97a0e2fcecfe70d7cfc1c9277318
CRs-Fixed: 1086498
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>