qcacld-3.0: Ignore LFR3 ROAM SYNCH indication if disconnect in progress
Currently, WMA is not checking whether the south bound disconnect
(deauth/disassoc from AP) is in progress before processing
LFR3 Roam synch indication which leads to FW out of sync with host.

Fix is to ignore LFR3 Roam Synch Indication if south bound disconnect
is in progress in host.

Change-Id: If3c78139cf1d1d0aa0f977bd2cbdca53c7b1a81c
CRs-Fixed: 2019702
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>