qcacld-3.0: Fix psoc's use after psoc destroy
psoc is allocated during start module and destroy in stop modules.
due to this if psoc is used after stop modules it cause crash.
psoc should be the first thing to get allocated and last thing
to get destroyed.

Fix it by calling psoc destroy at the ned of the wlan exit and
allocating in wlan start.

Change-Id: I969a5688751ddfad9ab9c3152e6af0680d2be605
CRs-Fixed: 2027351
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>