qcacld-3.0: Fix Kbuild to enable WLAN debugfs support
Currently, the CONFIG_WLAN_POWER_DEBUGFS and CONFIG_WLAN_DEBUGFS
are not enabled for statically loaded driver.
This results in missing subsystem power stats.
The reason for that is the enabling of this macro is conditionally
done under KERNEL_BUILD=0.
This commit enables the macro to be also
conditionally done under all values of KERNEL_BUILD

Change-Id: I2824fdbb438b37ee58a95a6ea889be5e7691bef5
CRs-Fixed: 2033114
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>