qcacld-3.0: Reject STA assoc on non-AP chan if DNSC enabled
It might happen that the STA-interface is in the process of
associating and before association is complete,
"Do_Not_Switch_Channel" is enabled on the AP-interface.
Reject STA association if "Do_Not_Switch_Channel" is enabled.
A check is performed in hdd_association_completion_handler() for
"Do_Not_Switch_Channel" in AP/GO vdev and STA association is
rejected accordingly.

Change-Id: I3f319a9e14c28add68f9809eeb86e782e58e72de
CRs-Fixed: 2036942
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>