qcacld-3.0: Reject connect if Do_Not_Switch_Channel set
In a case when a STA interface is existing and AP/GO comes up as the second
interface, it might happen that the 'scan list' of the existing STA
interface has not been refreshed yet.
It might be displaying a third-party-AP which is on a different
channel than our AP/GO-interface.
If the STA tries to connect to the third-party-AP, then the connection
is rejected if
1) Do_Not_Switch_Channel is set for the AP/GO vdev
 AND
2) The third-party-AP channel is on the same band but is different from the
AP/GO-interface channel.

If the third-party-AP and our AP/GO-interface channels are same OR in
different bands, then the connection is allowed to go through.

Change-Id: I940665ac38864cc8a16912c714412dd4c8d819d6
CRs-Fixed: 2036903
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>