qcacld-3.0: Fill rate and bw info for HT/VHT frames in monitor mode
In monitor mode, while filling the radiotap header, rate and bw
info for HT and VHT frames was not filled correctly due to which
packets in tcpdump are showing incorrect rates and bw for HT/VHT
frames.

Fill rate and bw info for HT/VHT frames correctly in monitor mode.

Change-Id: If15708101ac123fbdba141b04aad881d988e6b43
CRs-Fixed: 2034455
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>