qcacld-3.0: Call calbacks for hold req while removing it from queue
During SSR the hold req is not removed from queue and if HDD is
waiting for the response from WMA/FW, the HDD may timeout and due
to this shutdown ack is not sent to modem which results in modem
crash.

To avoid this, remove hold req from queue during ssr and call
their calback when SSR is in progress.

Change-Id: I462726c524cb15ff9aa88b051467f038ab1f4e67
CRs-Fixed: 2038200
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>