qcacld-3.0: Check Driver module status for sysfs read/write requests
When there is no interface is being used by the upper layer,interface
change timer kicks-in and closes all the modules. When there is a
sysfs read write request when the module is closed it tries to
access the contents of closed modules resulting in assert's or
crash.

To mitigate the issue validate the driver module status.

CRs-Fixed: 2052903
Change-Id: I2eec712ff823c7c2dc49f7d9d3c55d9166e3b0d9
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>