qcacmn: Change to address multiple logically deleted  node
In the existing code we check for the first instance of logically deleted
node and return to the caller, but there might be multiple nodes in logically
deleted state at a given instance.

Modify code to populate all the peers in logically deleted state in a list
and return the node with delete response pending from FW, to the caller.

Change-Id: I79ce99cc97404c2260ae18983709707ec6e8cdac
CRs-Fixed: 2058168
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>